Print contents of user bits in jltcdump as well
* Move print_user_bits() to new common_ltcdump.{c,h}
