+++ common_ltcdump.h
@@ -1,0 +1,10 @@
+#ifndef COMMON_LTCDUMP_H
+#define COMMON_LTCDUMP_H
+
+#include <stdio.h>
+#include <ltc.h>
+
+void print_user_bits(FILE *outfile, const LTCFrame *f);
+
+#endif
+/* vi:set ts=8 sts=2 sw=2: */
